// src/components/atomic-crm/providers/fakerest/dataGenerator/types.ts
import type {
  Company,
  Contact,
  ContactNote,
  Deal,
  DealNote,
  Sale,
  Region,
  Tag,
  Task,
} from "../../../types";
import type { ConfigurationContextValue } from "../../../root/ConfigurationContext";

export interface Db {
  companies: Company[];
  contacts: Contact[];
  contact_notes: ContactNote[];
  deals: Deal[];
  deal_notes: DealNote[];
  sales: Sale[];
  regions: Region[];
  tags: Tag[];
  tasks: Task[];
  configuration: Array<{ id: number; config: ConfigurationContextValue }>;
}